--- data/Black & White/Legendary Treasures/RC15.ts
@@ -1,71 +1,0 @@
-import { Card } from '../../../interfaces'
-import Set from '../Legendary Treasures'
-
-const card: Card = {
-	name: {
-		en: "Teddiursa",
-	},
-
-	illustrator: "kirisAki",
-	rarity: "Common",
-	category: "Pokemon",
-	set: Set,
-
-	dexId: [
-		216,
-	],
-
-	hp: 60,
-
-	types: [
-		"Colorless",
-	],
-
-	stage: "Basic",
-
-	attacks: [
-		{
-			cost: [
-				"Colorless",
-			],
-			name: {
-				en: "Honey Snack",
-			},
-			effect: {
-				en: "Heal 20 damage from this Pokémon.",
-			},
-
-		},
-		{
-			cost: [
-				"Colorless",
-				"Colorless",
-			],
-			name: {
-				en: "Scratch",
-			},
-
-			damage: 20,
-
-		},
-	],
-
-	weaknesses: [
-		{
-			type: "Fighting",
-			value: "×2"
-		},
-	],
-	retreat: 3,
-
-
-	description: {
-		en: "It lets honey soak into its paws so it can lick them all the time. Every set of paws tastes unique.",
-	},
-
-	thirdParty: {
-		cardmarket: 281133
-	}
-}
-
-export default card
