+++ data/Black & White/Radiant Collection/RC9.ts
@@ -1,0 +1,77 @@
+import { Card } from '../../../interfaces'
+import Set from '../Radiant Collection'
+
+const card: Card = {
+	name: {
+		en: "Kirlia",
+	},
+
+	illustrator: "Kagemaru Himeno",
+	rarity: "Common",
+	category: "Pokemon",
+	set: Set,
+
+	dexId: [
+		281,
+	],
+
+	hp: 70,
+
+	types: [
+		"Psychic",
+	],
+
+	evolveFrom: {
+		en: "Ralts",
+	},
+
+	stage: "Stage1",
+
+	attacks: [
+		{
+			cost: [
+				"Psychic",
+			],
+			name: {
+				en: "Tiptoe Step",
+			},
+			effect: {
+				en: "Draw a card for each Psychic Energy attached to this Pokémon.",
+			},
+
+		},
+		{
+			cost: [
+				"Psychic",
+				"Colorless",
+			],
+			name: {
+				en: "Double Slap",
+			},
+			effect: {
+				en: "Flip 2 coins. This attack does 20 damage times the number of heads.",
+			},
+			damage: 20,
+
+		},
+	],
+
+	weaknesses: [
+		{
+			type: "Psychic",
+			value: "×2"
+		},
+	],
+	retreat: 1,
+
+
+	description: {
+		en: "If its Trainer becomes happy, it overflows with energy, dancing joyously while spinning about.",
+	},
+
+	thirdParty: {
+		cardmarket: 281144
+	}
+}
+
+export default card
